refactor file upload component
